% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Store.R
\name{Search_Store}
\alias{Search_Store}
\title{Search_Store}
\usage{
Search_Store(PostalCode = 10025, Distance = 10)
}
\arguments{
\item{PostalCode}{Search for stores by postal code}

\item{Distance}{Radius of seaching area}
}
\description{
This function returns stores information for all the stores
within the radius of specified postal code.
}
\details{
This function provides stores information for all Best Buy stores
in the United States and Puerto Rico.
}
\keyword{BestBuy}
\keyword{Distance}
\keyword{PostalCode}
