% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_Trending_OpenBox.R
\name{Search_Trending_OpenBox}
\alias{Search_Trending_OpenBox}
\title{Search_Trending_OpenBox}
\usage{
Search_Trending_OpenBox(Category = "camera", Number_low = 10)
}
\arguments{
\item{Category}{Enter category}

\item{Number_low}{Number of products to display}
}
\description{
This function provides infomation of trending products with Open Box service
of a wanted category
}
\details{
This function allows the user to enter a specific category. It searches trending
products within the given category and returns specific number of trending products
with Open Box service.
}
\keyword{BestBuy}
\keyword{Category}
\keyword{OpenBox}
\keyword{TrendingProducts}
