% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Search_TV_Brand.R
\name{Search_TV_Brand}
\alias{Search_TV_Brand}
\title{Search_TV_Brand}
\usage{
Search_TV_Brand(Brand = "Sony", Rank = TRUE, Sort = TRUE)
}
\arguments{
\item{Brand}{Brand of the TV}

\item{Rank}{Best selling rank}

\item{Sort}{Sort by price}
}
\description{
This function a dashboard of characteristics of one specific TV.
}
\details{
This function allows the user to enter a brand name of TV
It will list all characteristics of that brand and
sort by the best-selling products over the last week.
}
\keyword{BestBuy}
\keyword{brand}
\keyword{price}
